requirements.txt
imapclient
beautifulsoup4
requests
readability-lxml
jinja2
python-dotenv
